Add/Update moref helper functions
We need to update `get_moref()` to support the way `zeep` is creating
and passing a ManagedObjectReference to the vCenter.

Since the attribute access changed from suds to zeep (before: .value and
._type; after: ._value_1 and .type), we introduce two helper functions
to be used everywhere a moref value/type needs to be extracted from a
ManagedObjectReference object or string. These functions are intended
to stop the actual SOAP library implementation to leak out of
oslo_vmware and in depending code. Having leaked abstractions in other
code make changes harder.

In that spirit, we also change all references inside of oslo_vmware to
using `get_moref_value()` and `get_moref_type()` instead of directly
accessing the attributes.

Change-Id: I208460c9293478f261176c25e8f0e2552b1dde69