Add serialize_object() helper function
This function converts a zeep object to a dictionary - an OrderedDict to
be precise. Some depending code had this implemented itself, but since
it's specific to the SOAP library, we pull it into oslo.vmware so we
don't leak specificas of the underlying SOAP library through the
oslo.vmware abstraction.

Change-Id: I6534e9e7994b110817315192e5ef7343e9b3743e