Make Service use zeep instead of suds
This is the biggest change in the migration from suds to zeep as it
touches the main object containing the SOAP client.

We define an interface for accessing the cookiejar, as it showed, that
it might be placed on different objects for different SOAP libraries. We
can now access it via '.client.cookiejar' on the 'Service' object.

The MemoryCache implementation used for suds gets removed, because zeep
comes with its own implementation. There are minimal differences in
behavior: one cannot set the timeout for a specific key anymore. But as
far as I have seen, this wasn't used anyways.

There is also no need for RequestsTransport anymore, as zeep uses
requests by default. We can also pass the parameters and the additional
LocalFileAdapter into the session used by requests as was done before.

While I can't say if the vSphere API still has the problems worked
around with 'ServiceMessagePlugin', we use two plugins for zeep doing
the same thing.

For keeping compatibility with code written for suds' 'Client' object,
we introduce a 'CompatibilityZeepClient' object. This is a zeep 'Client'
extended to contain the cookiejar attributes and a 'factory' attribute,
that behaves like the former 'factory' provided by suds. We also
overwrite the 'soap_url' in this client's '__init__()', because zeep
doesn't give us another way of changing it - it always takes the one
from the WSDL binding.

Change-Id: Icc781d025ef3987584dbe15dbb68554c22b9a2a1