Handle existing type on removeKey
The AddAnyTypeTypeAttributePlugin needs to keep the type of a removeKey,
if it's already set and only default to setting string or int otherwise.
Thre reason for this is, that the vSphere API expects the type of the
key of the array to be set as the type of the removeKey e.g. a
ManagedObjectReference.

Change-Id: I7aaf53b69e7f80716d316234213925be985179ce