SAP - Log bytes read/written
This patch adds the ability to log the bytes read/written as
well as percentage complete.

Unforatunately with vmware images, the total size and percentage
complete are almost always wrong.  vmware returns the total size
of the volume being backed up.  The amount of data from that volume
being read will almost always not be the same as the total volume size,
so the bytes read and percentage complete will never be accurate.
We need a new way to get the real data used inside the volume instead
of the volume(vmdk file) size.