Added Deserialize and compatible Serialize
To transport SOAP objects over other transport mechanisms,
such as oslo.messaging, we need a representation,
which allows us to reconstruct (deserialize) the
serialized object. This is commonly a dict, as done in
the pre-existing serialize_object function.

This change improves the function by:
1. Not serializing empty defaults making the result
   less verbose
2. Serialize the type by adding the "class" field to
   the dict, in cases of fields being of derived types
   of the default

For compatibility reasons, the latter is only active
when opting-in.

Additionally, the deserialize_object has been added,
which takes the dict generated by serialize_object
and creates a SOAP object out of it

Change-Id: Ie76b1e6940b5022563ce91d5692df589573704d0