Add .sleep() to avoid orphaned VMDKs
This is more of a hack, but I reproduced this in my local and
it fixes the orphaned VMDKs issues.
If the VMDK was not downloaded 100% and the caller calls close()
before the completion (i.e. when cinder backup fails and cinder
wants to disconnect from the backup device), it seems that VMware
still keeps the lock on the VMDK for a few seconds.
Deleting the VM right away would leave the VMDK orphaned,
because it still had a lock on it for a few seconds.
Delaying this by 1s-2s it still leaves orphaned VMDKs, but from 3s
onwards it seems to work. 5s is more of a safe measure.