VmdkReadHandle: gracefully shutdown and close the socket
Depending on the platform, the .close() does not guarantee that
the actual connection is closed right away.

From the Python docs:

close() releases the resource associated with a connection
but does not necessarily close the connection immediately.
If you want to close the connection in a timely fashion,
call shutdown() before close().

Change-Id: I9b916e8ef4458cba480ca3b74a9f3d70fded6303