Set HttpNfcLeaseSourceFile.create based on http_method
Currently we are hardcoding `create=True` which, according to the
docs, is enforcing `PUT` to be used for the upload.
The `create` parameter can be True if PUT should be used for
upload, otherwise POST.

A caller may specify `http_method` in the kwargs so we are
setting the correct value for `create` accordingly.

cinder already specifies `http_method='POST'`.